requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
json